Correction Bugs
Cette mise à jour permet de régler les bugs suivants:
*L'hébergement pouvait ne pas s'arrêter après la fermeture de l'app
*Il n'était pas possible d'ajouter un bot ou une extension sans provoquer un retour arrière dans l'app

+++ main_scripts/hosting.js
@@ -39,7 +39,9 @@ module.exports = {
         })
     },
     stopHosting(){
+        if (client){
         client.destroy()
+        }
         return {"success":true}
     }
 }
